final case

+++ OOP_Project_MusicShop/Main/Music.h
@@ -86,7 +86,7 @@ public:
 	int Cal_albumI_leaveing(int i) {
 		return (m_import[i - 1] - m_export[i - 1]);
 	}
-	int cal_price_AlbumI_quality(int i, int quality) {
+	double cal_price_AlbumI_quality(int i, int quality) {
 		return (m_album[i - 1].getPrice() * quality);
 	}
 	void Buy_AlbumI(int i, int quality) {
